coupons crud for users

+++ controllers/couponsController.js
@@ -1,24 +1,46 @@
 const Coupon = require('../models/coupons');
 const { infoLogger, errorLogger } = require("../logs/logs");
+const { userService } = require('../services/userService');
+
+const removeCode = (coupon) => {
+    if (coupon) {
+        coupon.code = undefined;
+    }
+    return coupon;
+}
+
 exports.couponsController = {
-    getCoupons(req, res) {
+    async getCoupons(req, res) {
         infoLogger.info("Get all Coupons");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
         Coupon.find({})
-            .then(coupon => {
+            .then(coupons => {
                 infoLogger.info("Success to Get all coupons");
-                res.json(coupon)
+                if (!isMod) {
+                    coupons = coupons.map((coupon) => removeCode(coupon));
+                }
+                res.json(coupons)
             })
             .catch(err => {
                 errorLogger.error(`Error getting the data from db:${err}`)
                 res.status(500).json({ "message": `Error Gets coupons ` });
             });
     },
-    getCouponDetails(req, res) {
+    async getCouponDetails(req, res) {
         infoLogger.info(`Get Coupon id:${req.params.id}`);
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
         Coupon.findOne({ _id: req.params.id })
             .then((coupon) => {
                 if (coupon) {
-                    res.json(coupon)
+                    if (!isMod)
+                        coupon = removeCoupon(coupon);
+                    res.json(removeCoupon(coupon))
                 }
                 else {
                     errorLogger.error("Wrong coupon id please enter correct id");
@@ -30,8 +52,17 @@ exports.couponsController = {
                 res.status(500).json({ "message": `Error getting coupon ` });
             });
     },
-    editCouponDetails(req, res) {
+    async editCouponDetails(req, res) {
         infoLogger.info("Updating a coupon");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         Coupon.updateOne({ _id: req.params.id }, req.body)
             .then((result) => {
                 if (result.matchedCount > 0) {
@@ -45,12 +76,28 @@ exports.couponsController = {
             })
             .catch((err) => res.status(400).json(err));
     },
-    addCoupon(req, res) {
+    async addCoupon(req, res) {
         infoLogger.info("Add a coupon");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         if (req.body.name && req.body.info &&
             req.body.code && req.body.imgUrl &&
             req.body.cost) {
-            const newCoupon = new Coupon(req.body);
+            const newCoupon = new Coupon({
+                name: req.body.name,
+                info: req.body.info,
+                code: req.body.code,
+                imgUrl: req.body.imgUrl,
+                cost: req.body.cost,
+                owned: req.body.owned ?? false,
+            });
             newCoupon.save()
                 .then(result => {
                     infoLogger.info(`Adding coupon  :${req.body.name} is successfully`);
@@ -66,8 +113,17 @@ exports.couponsController = {
             res.status(400).json({ "message": "Missing Parameters Please send all Parameters" });
         }
     },
-    deleteCoupon(req, res) {
+    async deleteCoupon(req, res) {
         infoLogger.info("Delete a coupon");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         Coupon.deleteOne({ _id: req.params.id })
             .then((result) => {
                 if (result.deletedCount > 0) {
